// packages/code-du-travail-frontend/src/modules/outils/preavis-retraite/steps/Result/utils/types.ts
export enum NoticeUsed {
  legal = "legal",
  agreementLabor = "agreementLabor",
  same = "same",
  none = "none",
}

export enum WarningType {
  noNoticeWithAgreement = "no_notice_with_agreement",
  noNoticeWithoutAgreement = "no_notice_without_agreement",
  departWithAgreement = "depart_with_agreement",
  miseWithAgreement = "mise_with_agreement",
  departWithoutAgreement = "depart_without_agreement",
  miseWithoutAgreement = "mise_without_agreement",
  miseWithoutCollectiveAgreement = "mise_without_collective_agreement",
  departWithoutCollectiveAgreement = "depart_without_collective_agreement",
}